feat: Implement Carousel functionality with draggable items
- Added CarouselItem class to manage individual carousel items.
- Introduced DraggableCarouselItem class for handling drag events and animations.
- Created CarouselCardItem class for card-specific behavior and selection logic.
- Developed CarouselState class to maintain the state of the carousel, including dragging status and selected item.
- Implemented position tracking and animations using GSAP for smooth interactions.
